# contracts/backend-api/1.0.0/openapi.json
{
  "components": {
    "responses": {
      "ContractError": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        },
        "description": "使用稳定错误代码表示的失败。"
      }
    },
    "schemas": {
...
      "ApiErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "schemaVersion": {
            "const": "1.0.0"
          }
        },
        "required": [
          "schemaVersion",
          "error"
        ],
        "type": "object"
      },
...
      "Body_publish": {
        "properties": {
          "archive": {
            "format": "binary",
            "title": "Archive",
            "type": "string"
          },
          "metadata": {
            "title": "Metadata",
            "type": "string"
          }
        },
        "required": [
          "archive",
          "metadata"
        ],
        "title": "Body_publish",
        "type": "object"
      },
      "Body_update_publication_information_with_cover": {
        "properties": {
          "cover": {
            "format": "binary",
            "title": "Cover",
            "type": "string"
          },
          "information": {
            "title": "Information",
            "type": "string"
          }
        },
        "required": [
          "information",
          "cover"
        ],
        "title": "Body_update_publication_information_with_cover",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "siteSession": {
        "in": "header",
        "name": "X-PbDH-Session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "PbDH Platform 的身份、云文档、托管媒体和资源市场 HTTP 合约。领域 payload 仍由各自 Platform Contract 管理。",
    "title": "PbDH Platform API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/publications": {
      "get": {
        "operationId": "list_publications",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "templateId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Templateid"
            }
          },
          {
            "in": "query",
            "name": "targetSystemPackageId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Targetsystempackageid"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "authorAccountId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authoraccountid"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "recent",
              "enum": [
                "relevance",
                "recent",
                "title"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 60,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Publications",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [],
        "summary": "List Publications",
        "tags": [
          "publications"
        ]
      },
      "post": {
        "operationId": "publish",
        "parameters": [
          {
            "in": "header",
            "name": "X-PbDH-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pbdh-Session"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_publish"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Publish",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "siteSession": []
          }
        ],
        "summary": "Publish",
        "tags": [
          "publications"
        ]
      }
    },
    "/api/publications/manageable": {
      "get": {
        "operationId": "list_manageable_publications",
        "parameters": [
          {
            "in": "header",
            "name": "X-PbDH-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pbdh-Session"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Manageable Publications",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "siteSession": []
          }
        ],
        "summary": "List Manageable Publications",
        "tags": [
          "publications"
        ]
      }
    },
    "/api/publications/{publication_id}": {
      "delete": {
        "operationId": "delete_publication",
        "parameters": [
          {
            "in": "path",
            "name": "publication_id",
            "required": true,
            "schema": {
              "title": "Publication Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-PbDH-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pbdh-Session"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "siteSession": []
          }
        ],
        "summary": "Delete Publication",
        "tags": [
          "publications"
        ]
      },
      "get": {
        "operationId": "get_publication",
        "parameters": [
          {
            "in": "path",
            "name": "publication_id",
            "required": true,
            "schema": {
              "title": "Publication Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Publication",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [],
        "summary": "Get Publication",
        "tags": [
          "publications"
        ]
      }
    },
    "/api/publications/{publication_id}/download": {
      "get": {
        "operationId": "download_publication",
        "parameters": [
          {
            "in": "path",
            "name": "publication_id",
            "required": true,
            "schema": {
              "title": "Publication Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-PbDH-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pbdh-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.pbdh.resource-package+zip": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Resource Package 便携归档。"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": [],
            "siteSession": []
          }
        ],
        "summary": "Download Publication",
        "tags": [
          "publications"
        ]
      }
    },
    "/api/publications/{publication_id}/information": {
      "patch": {
        "operationId": "update_publication_information",
        "parameters": [
          {
            "in": "path",
            "name": "publication_id",
            "required": true,
            "schema": {
              "title": "Publication Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-PbDH-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pbdh-Session"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Publication Information",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "siteSession": []
          }
        ],
        "summary": "Update Publication Information",
        "tags": [
          "publications"
        ]
      }
    },
    "/api/publications/{publication_id}/information-with-cover": {
      "patch": {
        "operationId": "update_publication_information_with_cover",
        "parameters": [
          {
            "in": "path",
            "name": "publication_id",
            "required": true,
            "schema": {
              "title": "Publication Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-PbDH-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pbdh-Session"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_publication_information_with_cover"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Publication Information With Cover",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ContractError"
          },
          "default": {
            "$ref": "#/components/responses/ContractError"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "siteSession": []
          }
        ],
        "summary": "Update Publication Information With Cover",
        "tags": [
          "publications"
        ]
      }
    },
...
  },
...
}